requirements.txt
requests
sentence-transformers
faiss-cpu
ollama
numpy
